Update template
Thanks for the corrections!  It builds for me and I have learned since making this PR that instead of adding more variables or flags and, therefore, complexity one should go in the other direction toward simplicity and let the build_style take over. At least that seems to be the case if build_style=cmake as I ran into problems with a python pkg using python build style.

Also "ew" is spelled "eww" and for emphasis  "ewwa" or more ws like "ewwwwwww". ;)